SPR Added to results and displayed
Correctly added the results from the spr to the results table and was able to call it back and display it back to the user. Had to combine 2 sections into 1. E.g county - partyname as this allowed me to call it back and strip them to easily get it back into a dictionary.